# openapi.yaml
openapi: 3.0.3
info:
  title: Cyber Storefront API
  version: "1.0.0"
  description: |
    **Cyber** ონლაინ მაღაზიის API: კატალოგი, ავტორიზაცია, კალათა, ჩექაუთი და პროფილი.

    ---

    # 🚀 სწრაფი დაწყება — სცადე აქვე, 4 ნაბიჯად

    ამ გვერდზე ყველა ენდპოინტის გატესტვა შეგიძლია, კოდის დაწერის გარეშე.

    ### ნაბიჯი 1 — შექმენი ანგარიში

    ჩამოდი `POST /auth/register`-მდე → დააჭირე **`Try it out`** (მარჯვნივ)
    → შეცვალე მონაცემები → **`Execute`**

    ```json
    { "name": "Test User", "email": "შენი@example.com", "password": "Password1" }
    ```

    ### ნაბიჯი 2 — დააკოპირე ტოკენი

    პასუხში (Response body) ნახავ:

    ```json
    {
      "user": { "id": "...", "name": "Test User", "email": "..." },
      "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOi..."
    }
    ```

    დააკოპირე **`accessToken`-ის მნიშვნელობა** — ბრჭყალების გარეშე.

    ### ნაბიჯი 3 — დააჭირე `Authorize` 🔓

    ამ გვერდის **ზედა მარჯვენა კუთხეში** არის ღილაკი **`Authorize`** (ბოქლომის ხატულით).

    1. დააჭირე
    2. ველში `Value` ჩასვი დაკოპირებული ტოკენი
       *(მხოლოდ ტოკენი — სიტყვა `Bearer` ჩაწერა არ გჭირდება, Swagger თავად დაამატებს)*
    3. **`Authorize`** → **`Close`**

    ბოქლომები 🔒 დაიკეტება — ესე იგი ავტორიზებული ხარ.

    ### ნაბიჯი 4 — სცადე დაცული ენდპოინტი

    ჩამოდი `GET /auth/me`-მდე → **`Try it out`** → **`Execute`**

    ✅ დაბრუნდება შენი მომხმარებლის მონაცემები.

    **სცადე საწინააღმდეგოც:** დააჭირე `Authorize` → **`Logout`** → ისევ გაუშვი
...

    ---

...

...
servers:
...
paths:
...
  /cart:
    get:
      tags: [Cart]
      summary: ჩემი კალათა
      description: 🔒 აბრუნებს პოზიციებს, საერთო რაოდენობას (`totalQty`) და ჯამს (`subtotal`).
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      tags: [Cart]
      summary: კალათის გასუფთავება
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /cart/items:
    post:
      tags: [Cart]
      summary: კალათაში დამატება
      description: |
        🔒 თუ პროდუქტი კალათაში უკვე არის, **რაოდენობა ემატება** (ახალი პოზიცია არ იქმნება)
        — პასუხი მაშინ `200`-ია, ახალზე `201`.

        ყველა ცვლილება აბრუნებს **მთლიან განახლებულ კალათას** — ცალკე `GET /cart` არ გჭირდება.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId: { type: string, description: "პროდუქტის `id` (არა slug)" }
                qty: { type: integer, minimum: 1, maximum: 99, default: 1 }
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "201": { $ref: "#/components/responses/Cart" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: პროდუქტი არ არსებობს (`PRODUCT_NOT_FOUND`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409": { $ref: "#/components/responses/NotEnoughStock" }
        "422":
          $ref: "#/components/responses/Validation"
  /cart/items/{id}:
    parameters:
      - { name: id, in: path, required: true, description: "კალათის პოზიციის `id` (`items[].id`, არა productId)", schema: { type: string } }
    patch:
      tags: [Cart]
      summary: რაოდენობის შეცვლა (+ / −)
      description: |
        🔒 აყენებს **ზუსტ** რაოდენობას. სტეპერზე `+` → `qty + 1`, `−` → `qty - 1`.
        `0` დაუშვებელია — პოზიციის მოსაშორებლად გამოიყენე `DELETE`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qty]
              properties:
                qty: { type: integer, minimum: 1, maximum: 99, example: 2 }
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/CartItemNotFound" }
        "409": { $ref: "#/components/responses/NotEnoughStock" }
        "422":
          $ref: "#/components/responses/Validation"
    delete:
      tags: [Cart]
      summary: პოზიციის წაშლა
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/Cart" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/CartItemNotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: ტოკენი არ არის, არასწორია ან ვადაგასულია (`UNAUTHORIZED` / `TOKEN_EXPIRED`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Cart:
      description: განახლებული კალათა
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Cart" }
    NotEnoughStock:
      description: მარაგი არ კმარა (`OUT_OF_STOCK` / `INSUFFICIENT_STOCK`); `available` — რამდენია მარაგში
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string, example: "Only 3 left in stock" }
              code: { type: string, example: INSUFFICIENT_STOCK }
              available: { type: integer, example: 3 }
    CartItemNotFound:
      description: პოზიცია არ მოიძებნა (`CART_ITEM_NOT_FOUND`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Validation:
      description: ვალიდაციის შეცდომა (`VALIDATION_ERROR`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
  schemas:
...
    CartItem:
      type: object
      properties:
        id: { type: string, description: "პოზიციის id — PATCH/DELETE-ისთვის" }
        productId: { type: string }
        qty: { type: integer, example: 2 }
        lineTotal: { type: number, example: 199.98 }
        product:
          type: object
          properties:
            id: { type: string }
            slug: { type: string }
            title: { type: string }
            brand: { type: string }
            image: { type: string }
            price: { type: number, example: 99.99 }
            oldPrice: { type: number, nullable: true }
            currency: { type: string, example: GEL }
            stock: { type: integer, example: 12 }
            inStock: { type: boolean }
    Cart:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/CartItem" } }
        totalQty: { type: integer, example: 3, description: "ჰედერის კალათის ბეჯისთვის" }
        subtotal: { type: number, example: 299.97 }
        currency: { type: string, example: GEL }
...
    Error:
      type: object
      properties:
        message: { type: string, example: "Invalid email or password" }
        code: { type: string, example: "INVALID_CREDENTIALS" }
    ValidationError:
      type: object
      properties:
        message: { type: string, example: "Validation failed" }
        code: { type: string, example: "VALIDATION_ERROR" }
        errors:
          type: object
          additionalProperties: { type: string }
          example: { email: "Enter a valid email address", password: "Password must be at least 8 characters" }